remove vendor, clean up module files and Makefile
It's better not to have the vendor directory; it was needed before
modules worked well in Go. These days, the `go.mod` and `go.sum` files
nail down the exact versions of each dependency as well as the vendor
directory did. There are many modern Go projects, such as
docker/compose, that don't have a vendor directory.

So don't run `go vendor`, or if you do, don't git-add the files that
are downloaded.

Another advantage of removing the vendor directory is that the repo is
only 9% of its former size.

Another thing I did as part of this commit is to remove all the
`require` lines in `go.mod` and then ran `go tidy`. This repopulated the
require lines but with the latest versions. This may fix #378 but I'm
not sure (because I still see `gopkg.in/yaml.v2 v2.4.0 // indirect`
in `go.mod`.