Add debug logging to gRPC entrypoints
This additional logging may help with debugging lightwalletd clients as
it may isolate the problem to the client or server (lightwalletd) side.

To see entrypoint logging, you must start lightwalletd with the option
--log-level 5 (which is "debug"; the default is 4, which is "info")).

This option shouldn't be used in normal operation, because clients could
flood the log, causing crashes due to running out of disk space. Only
use this when actually debugging or testing.

For most gRPCs, also log return values if the log-level is 6 ("trace").
Return value logging isn't done if there is an error (it's assumed that
the client will log something) or if the output might be very large (we
don't want to overrun the log file).

Enabling trace (--log-level 6) enables everything below that level,
so entry logging is enabled as well.

(cherry picked from commit 7ec4d1c67c3225698fa32d375df20915512f860b)